tracking keypresses, problems with polyphony/note release
