+++ src/components/Keyboard.jsx
@@ -4,10 +4,10 @@ import Key from "./Key";
 
 //hooks
 import useAudioContext from "../useAudioContext";
+import { useState, useRef, useEffect } from "react";
 
 //styles
 import "../styles/component-styles/keyboard.scss";
-import { useState } from "react";
 
 export default function Keyboard() {
   const { 
@@ -21,31 +21,56 @@ export default function Keyboard() {
     setOctaveModifier
    } = useAudioContext();
 
+  const [keysPressed, setKeysPressed] = useState([]);
+
+  const handleKeyDown = (event) => {
+    const update = [...keysPressed, (event.key).toLowerCase()];
+    setKeysPressed(update);
+  }
+
+  const handleKeyUp = (event) => {
+    const update = keysPressed.filter(key => key !== (event.key).toLowerCase());
+    setKeysPressed(update);
+  }
+
+  const ref = useRef(null);
+
+  useEffect(() => {
+    ref.current.focus();
+  }, []);
+
+  let i = -1;
+
   const pianoKeys = noteFreq.map((keys, index) => {
     const keyList = Object.entries(keys);
     const keyboardInputs = ['a', 'w', 's', 'd', 'r', 'f', 't', 'g', 'h', 'u', 'j', 'i', 'k', 'o', 'l', ';'];
-
-    let i = -1;
-
+    
     return keyList.map((key) => {
       i++;
       return (
         <Key 
-          key={key[0]}
+          key={`${key[0]}${key[1]}`}
           note={key[0]}
           octave={index}
           freq={key[1]}
           notePressed={notePressed}
           noteReleased={noteReleased}
+          keyDown={handleKeyDown}
+          keyUp={handleKeyUp}
           input={keyboardInputs[i]}
+          keysPressed={keysPressed}
           whiteKey={key[0].length === 2 ? false : true}
         />
-      )
+        )
+      })
     })
-  })
 
   return (
-    <div className="keyboard">
+    <div className="keyboard"
+      ref={ref}
+      tabIndex={-1}
+      onKeyDown={e => handleKeyDown(e)}
+      onKeyUp={e => handleKeyUp(e)}>
       <Controls 
         sliders={sliders} 
         onChange={changeSliders}
